chore: Add unique reference id to close channel
Before we were reusing the same reference id of the channel.
